+++ src/main/java/com/getindata/connectors/http/internal/sink/HttpSinkWriter.java
@@ -3,6 +3,9 @@ package com.getindata.connectors.http.internal.sink;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
+import java.util.Properties;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 import java.util.function.Consumer;
 
 import lombok.extern.slf4j.Slf4j;
@@ -11,8 +14,11 @@ import org.apache.flink.connector.base.sink.writer.AsyncSinkWriter;
 import org.apache.flink.connector.base.sink.writer.BufferedRequestState;
 import org.apache.flink.connector.base.sink.writer.ElementConverter;
 import org.apache.flink.metrics.Counter;
+import org.apache.flink.util.concurrent.ExecutorThreadFactory;
 
 import com.getindata.connectors.http.internal.SinkHttpClient;
+import com.getindata.connectors.http.internal.config.HttpConnectorConfigConstants;
+import com.getindata.connectors.http.internal.utils.ThreadUtils;
 
 /**
  * Sink writer created by {@link com.getindata.connectors.http.HttpSink} to write to an HTTP
@@ -25,6 +31,13 @@ import com.getindata.connectors.http.internal.SinkHttpClient;
  */
 @Slf4j
 public class HttpSinkWriter<InputT> extends AsyncSinkWriter<InputT, HttpSinkRequestEntry> {
+
+    private static final String HTTP_SINK_WRITER_THREAD_POOL_SIZE = "4";
+
+    /**
+     * Thread pool to handle HTTP response from HTTP client.
+     */
+    private final ExecutorService sinkWriterThreadPool;
 
     private final String endpointUrl;
 
@@ -43,7 +56,9 @@ public class HttpSinkWriter<InputT> extends AsyncSinkWriter<InputT, HttpSinkRequ
             long maxRecordSizeInBytes,
             String endpointUrl,
             SinkHttpClient sinkHttpClient,
-            Collection<BufferedRequestState<HttpSinkRequestEntry>> bufferedRequestStates) {
+            Collection<BufferedRequestState<HttpSinkRequestEntry>> bufferedRequestStates,
+            Properties properties) {
+
         super(elementConverter, context, maxBatchSize, maxInFlightRequests, maxBufferedRequests,
             maxBatchSizeInBytes, maxTimeInBufferMS, maxRecordSizeInBytes, bufferedRequestStates);
         this.endpointUrl = endpointUrl;
@@ -51,6 +66,17 @@ public class HttpSinkWriter<InputT> extends AsyncSinkWriter<InputT, HttpSinkRequ
 
         var metrics = context.metricGroup();
         this.numRecordsSendErrorsCounter = metrics.getNumRecordsSendErrorsCounter();
+
+        int sinkWriterThreadPollSize = Integer.parseInt(properties.getProperty(
+            HttpConnectorConfigConstants.SINK_HTTP_WRITER_THREAD_POOL_SIZE,
+            HTTP_SINK_WRITER_THREAD_POOL_SIZE
+        ));
+
+        this.sinkWriterThreadPool =
+            Executors.newFixedThreadPool(
+                sinkWriterThreadPollSize,
+                new ExecutorThreadFactory(
+                    "http-sink-writer-worker", ThreadUtils.LOGGING_EXCEPTION_HANDLER));
     }
 
     // TODO: Reintroduce retries by adding backoff policy
@@ -59,7 +85,7 @@ public class HttpSinkWriter<InputT> extends AsyncSinkWriter<InputT, HttpSinkRequ
             List<HttpSinkRequestEntry> requestEntries,
             Consumer<List<HttpSinkRequestEntry>> requestResult) {
         var future = sinkHttpClient.putRequests(requestEntries, endpointUrl);
-        future.whenComplete((response, err) -> {
+        future.whenCompleteAsync((response, err) -> {
             if (err != null) {
                 int failedRequestsNumber = requestEntries.size();
                 log.error(
@@ -89,11 +115,17 @@ public class HttpSinkWriter<InputT> extends AsyncSinkWriter<InputT, HttpSinkRequ
                 //}
             }
             requestResult.accept(Collections.emptyList());
-        });
+        }, sinkWriterThreadPool);
     }
 
     @Override
     protected long getSizeInBytes(HttpSinkRequestEntry s) {
         return s.getSizeInBytes();
+    }
+
+    @Override
+    public void close() {
+        sinkWriterThreadPool.shutdownNow();
+        super.close();
     }
 }