+++ src/test/java/com/getindata/connectors/http/internal/table/lookup/AsyncHttpTableLookupFunctionTest.java
@@ -10,6 +10,9 @@ import java.util.Set;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.CyclicBarrier;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicInteger;
 
 import lombok.extern.slf4j.Slf4j;
 import org.apache.flink.table.data.RowData;
@@ -45,6 +48,8 @@ class AsyncHttpTableLookupFunctionTest {
 
     @BeforeEach
     public void setUp() throws Exception {
+        when(decorate.getOptions()).thenReturn(HttpLookupConfig.builder().build());
+
         barrier = new CyclicBarrier(rowKeys.length);
         threadNames = Collections.synchronizedSet(new HashSet<>());
         asyncFunction = new AsyncHttpTableLookupFunction(decorate);
@@ -69,7 +74,11 @@ class AsyncHttpTableLookupFunctionTest {
                 });
         }
 
-        latch.await();
+        assertThat(latch.await(3, TimeUnit.SECONDS))
+            .withFailMessage(
+                "Future complete in AsyncHttpTableLookupFunction was not called"
+                    + " for at lest one event.")
+            .isEqualTo(true);
 
         assertThat(result.size()).isEqualTo(rowKeys.length);
         assertThat(threadNames.size()).isEqualTo(rowKeys.length);
@@ -83,6 +92,7 @@ class AsyncHttpTableLookupFunctionTest {
         final List<RowData> result = Collections.synchronizedList(new ArrayList<>());
 
         CountDownLatch latch = new CountDownLatch(rowKeys.length);
+        AtomicBoolean wasException = new AtomicBoolean(false);
 
         for (int key : rowKeys) {
             CompletableFuture<Collection<RowData>> future = new CompletableFuture<>();
@@ -99,12 +109,67 @@ class AsyncHttpTableLookupFunctionTest {
                     })
                 .exceptionally(
                     throwable -> {
+                        wasException.set(true);
                         latch.countDown();
                         return Collections.emptyList();
                     });
         }
 
-        latch.await();
+        assertThat(latch.await(3, TimeUnit.SECONDS))
+            .withFailMessage(
+                "Future complete in AsyncHttpTableLookupFunction was not called"
+                    + " for at lest one event.")
+            .isEqualTo(true);
+
+        assertThat(wasException).isTrue();
+
+        // -1 since one will have an exception
+        assertThat(result.size()).isEqualTo(rowKeys.length - 1);
+        assertThat(threadNames.size()).isEqualTo(rowKeys.length);
+        verify(decorate, times(rowKeys.length)).lookupByKeys(any());
+    }
+
+    @Test
+    void shouldHandleEmptyOptionalResult() throws InterruptedException {
+        mockPollingWithEmptyOptional();
+
+        final List<RowData> result = Collections.synchronizedList(new ArrayList<>());
+
+        CountDownLatch latch = new CountDownLatch(rowKeys.length);
+        AtomicInteger completeCount = new AtomicInteger(0);
+
+        for (int key : rowKeys) {
+            CompletableFuture<Collection<RowData>> future = new CompletableFuture<>();
+            asyncFunction.eval(future, key);
+            future
+                .whenComplete(
+                    (rs, t) -> {
+                        if (t != null) {
+                            log.error(t.getMessage(), t);
+                        }
+
+                        completeCount.incrementAndGet();
+                        result.addAll(rs);
+                        latch.countDown();
+                    });
+        }
+
+        assertThat(latch.await(3, TimeUnit.SECONDS))
+            .withFailMessage(
+                "Future complete in AsyncHttpTableLookupFunction was not called"
+                    + " for at lest one event.")
+            .isEqualTo(true);
+
+        assertThat(completeCount.get())
+            .withFailMessage(
+                "Future complete in AsyncHttpTableLookupFunction was not called"
+                    + " for at lest one event.")
+            .isEqualTo(rowKeys.length);
+
+        // -1 since one will have one empty result.
+        assertThat(result.size()).isEqualTo(rowKeys.length - 1);
+        assertThat(threadNames.size()).isEqualTo(rowKeys.length);
+        verify(decorate, times(rowKeys.length)).lookupByKeys(any());
     }
 
     private void mockPolling() {
@@ -129,6 +194,21 @@ class AsyncHttpTableLookupFunctionTest {
                     Integer argument = (Integer) ((Object[]) invocationOnMock.getArgument(0))[0];
                     if (argument == 12) {
                         throw new RuntimeException("Exception On problematic item");
+                    }
+                    return Optional.of(buildGenericRowData(Collections.singletonList(argument)));
+                });
+    }
+
+    private void mockPollingWithEmptyOptional() {
+        when(decorate.lookupByKeys(any()))
+            .thenAnswer(
+                invocationOnMock -> {
+                    threadNames.add(Thread.currentThread().getName());
+                    // make sure we pile up all keyRows on threads
+                    barrier.await();
+                    Integer argument = (Integer) ((Object[]) invocationOnMock.getArgument(0))[0];
+                    if (argument == 12) {
+                        return Optional.empty();
                     }
                     return Optional.of(buildGenericRowData(Collections.singletonList(argument)));
                 });